refactor: DokiHome into DokiEggView and DokiView subcomponents
